[Silver I] Title: 직각삼각형, Time: 1184 ms, Memory: 2044 KB -BaekjoonHub

+++ 백준/Silver/1711. 직각삼각형/직각삼각형.cc
@@ -1,12 +1,12 @@
 #include <bits/stdc++.h>
 #define FASTIO ios_base::sync_with_stdio(false), cin.tie(nullptr), cout.tie(nullptr);
 using namespace std;
-#define int long long int
+#define ll long long int
 
 int n;
 int ans = 0;
-pair<int, int> xy[1500];
-int dist[3];
+pair<ll, ll> xy[1500];
+ll dist[3];
 
 signed main()
 {
@@ -24,9 +24,9 @@ signed main()
                 auto [x1, y1] = xy[i];
                 auto [x2, y2] = xy[j];
                 auto [x3, y3] = xy[k];
-                dist[0] = (x2-x1)*(x2-x1) + (y2-y1)*(y2-y1);
-                dist[1] = (x3-x2)*(x3-x2) + (y3-y2)*(y3-y2);
-                dist[2] = (x3-x1)*(x3-x1) + (y3-y1)*(y3-y1);
+                dist[0] = ((ll)x2-x1)*(x2-x1) + ((ll)y2-y1)*(y2-y1);
+                dist[1] = ((ll)x3-x2)*(x3-x2) + ((ll)y3-y2)*(y3-y2);
+                dist[2] = ((ll)x3-x1)*(x3-x1) + ((ll)y3-y1)*(y3-y1);
                 if(2*max(dist[0], max(dist[1], dist[2])) == dist[0]+dist[1]+dist[2]) ans++;
             }
         }
